requirements.txt
blessings==1.6
bpython==0.14.1
chardet==2.3.0
curtsies==0.1.19
Flask==0.10.1
greenlet==0.4.7
itsdangerous==0.24
Jinja2==2.7.3
MarkupSafe==0.23
Pygments==2.0.2
python-memcached==1.54
requests==2.7.0
six==1.9.0
Werkzeug==0.10.4
